+++ apps/dashboard/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @stackframe/stack-dashboard
 
+## 2.5.34
+
+### Patch Changes
+
+- Internationalization
+- Updated dependencies
+  - @stackframe/stack-shared@2.5.34
+  - @stackframe/stack@2.5.34
+  - @stackframe/stack-emails@2.5.34
+  - @stackframe/stack-ui@2.5.34
+
 ## 2.5.33
 
 ### Patch Changes